requirements.txt
       google-generativeai
       Flask